Trash: the legacy bin deleted whole — the app claims the frozen id
The side-by-side demonstration is over: the App Framework Trash is
THE Recycle Bin. The app takes the frozen desktop-mode-recycle-bin
identifier (AGENTS.md), which is what lets everything bound to it
survive with zero retargeting: desktop shortcuts, dock placement and
the Apps & Plugins row, the drag-to-trash drop targets (the window
root keeps the data-os-recycle-bin-root hook; the closed dock tile
keeps its selector), the desktop-theme icon slots, dock-peek, and
the closed-window empty/full art swap (icon-state.ts, untouched).
It also inherits the legacy rail furniture verbatim: dock control,
order 40, placeable, no desktop icon — pinned by the same
desktopObjectTitles / nativeWindowDockOrder suites, now pointed at
the app registration.

Deleted: the 1,305-line imperative bundle (src/recycle-bin/index.ts),
its 198-line REST client (rest.ts — the wire types move to a pure
types.ts leaf), the vite target and build script, the script-handle
registration, and window.php's registration/template/localize
(2,155 → 187 lines: it keeps the tile art, the capability gate, the
shell-config seed, and the stylesheet enqueue — the drag-to-trash
highlight must exist at boot, not window-open). The store, capture,
realtime and REST routes are untouched; the app and third parties
keep using them.

Removed public surface — openstation_recycle_bin_window_args and
_template_html (openstation_app_manifest covers the former; there is
no server template to filter) — is written up in
docs/migration-recycle-bin-app.md and flagged in hooks-reference.
Tests: the template/UI suites for the deleted window go
(media-tab, empty-state kses, selection-identity, empty-chrome,
read-silence — their behavior lives in trashApp.php + trash.test.ts);
store, empty-loop, pin-to-desktop and icon-state suites stay green;
lazyWindowConfigPriority and deferredWindowStyles follow the moved
enqueue. i18n handle map repointed at the app's companion handle.
