+++ docs/README.md
@@ -37,6 +37,7 @@ If you are **building a plugin** that interacts with the desktop shell — opens
 29. **[The App Framework — a window in one PHP file](./app-framework.md)** — *Experimental.* `.os.php` apps: declare a whole window (title, size, title-bar buttons, ⋯ rows, tabs, state, actions, view) in PHP, with `<os-*>` components and a small attribute vocabulary (`os-action`, `os-bind`, `os-poll`, `os-prop-*`, …), and ship no JavaScript — or add an `.os.ts` client view (`defineApp()`) for the interactions that must be instant. Covers the state cycle, the `$os` data-access layer that keeps the core WordPress-free, the standalone host, and the client runtime. Code Blue (`apps/code-blue/`) is built on it. See also [examples/os-app.md](./examples/os-app.md).
 30. **[Migration — Code Blue becomes an App Framework app](./migration-code-blue-app.md)** — the `desktop-mode/v1/code-blue/*` REST routes and the `openstation_code_blue_{window_args,icon_args,template_html}` filters are gone; the window, its id, its gate and its other hooks are unchanged. Read if you called those routes or filters.
 31. **[Migration — WP Explorer becomes the `my-wordpress` app](./migration-wp-explorer-app.md)** — the legacy `desktop-mode-my-wordpress` native window, its bundle and the `wp.os.myWordpress` API are gone; the app carries the name, the launcher and every surface (footprint included). Read if you opened, configured or extended the legacy window.
+32. **[Migration — the Recycle Bin becomes an App Framework app](./migration-recycle-bin-app.md)** — the `openstation_recycle_bin_{window_args,template_html}` filters, the legacy bundle and its config global are gone; the frozen window id, the store/capture/realtime hooks, the REST routes and the JS columns filter are unchanged. Read if you filtered the bin's registration or template.
 32. **[The Living Tree — algorithm definition](./living-tree-algorithm.md)** — *Experimental.* The full normative spec for the `wp-living-tree` canvas wallpaper: WordPress emits hormones, the biology (Space Colonization) decides geometry inside age-bounded morphological constraints. Read before touching any part of the wallpaper.
 33. **[Multisite](./multisite.md)** — *Experimental.* What the shell does on a network: the network admin, the Network Admin dock tile, and per-site session scoping. Read before touching anything cross-site — it opens with the constraint that decides the rest, which is that WordPress refuses to be framed cross-origin, so nothing leaving the current site can be a window.
