+++ includes/recycle-bin/window.php
@@ -1,21 +1,17 @@
 <?php
 /**
- * OpenStation — Recycle Bin: window registration.
+ * OpenStation — Recycle Bin: tile art, gate, shell config.
  *
- * Native window with id `desktop-mode-recycle-bin`, pinned to the dock.
- * Like the code editor, the template body is a static skeleton that the
- * JS bundle enhances on first open — the table is populated from the
- * REST list endpoint at render time.
- *
- * The bin lands on the dock and nowhere else. It used to also register
- * a wallpaper icon, which put the same target on two surfaces at once
- * and made the desktop something the shell furnished rather than
- * something the user did. That is a default, not a rule: the tile is
- * `placeable`, so the wallpaper is one pick away in Apps & Plugins.
- *
- * The registration is filterable via `openstation_recycle_bin_window_args`
- * so a plugin can swap the icon, change the dimensions, or restrict who
- * sees the bin without touching this file.
+ * The bin WINDOW is an App Framework app (`apps/trash/trash.os.php`)
+ * that claims the frozen `desktop-mode-recycle-bin` id, so every
+ * shortcut, dock placement and drop target keeps working. This file
+ * keeps what the shell needs regardless of the window: the two bin
+ * drawings (the dock tile's empty/full swap — `icon-state.ts` — and
+ * the app's own `ctx.host.setIcon()` swap read the same pair), the
+ * capability gate, the shell-config seed that makes the tile correct
+ * on the very first paint, and the stylesheet enqueue (the
+ * drag-to-trash drop-target highlight must exist at boot, not at
+ * window-open).
  *
  * @package OpenStation
  */
@@ -23,7 +19,7 @@
 defined( 'ABSPATH' ) || exit;
 
 /**
- * The bin SVG, used by the window icon and its dock tile.
+ * The bin SVG, used by the app window's icon and its dock tile.
  *
  * The bin used to be `dashicons-trash`, which worked but wore the
  * wrong clothes. Dashicons are WP core's icon set: solid fills on a
@@ -50,9 +46,7 @@ defined( 'ABSPATH' ) || exit;
  * match their neighbours rather than this icon.
  *
  * The bin has two states. Empty is the vessel on its own; full adds
- * three crumpled balls inside it and knocks the lid askew. See
- * {@link openstation_recycle_bin_icon_svg()} for why the lid only
- * moves 8 degrees.
+ * three crumpled balls inside it and knocks the lid askew.
  *
  * @param bool $full Whether to draw the bin holding something.
  * @return string Raw `<svg>` markup.
@@ -129,113 +123,11 @@ function openstation_recycle_bin_icon_uris() {
 }
 
 /**
- * Echoes the recycle bin window's template body.
- *
- * The shell wraps this in `<template id="os-native-window-desktop-mode-recycle-bin">`
- * and clones it into the window body BEFORE the JS render callback runs.
- * The `data-os-recycle-bin-*` hooks below are the contract the JS
- * relies on — keep them intact (or rename via the filter) when
- * customizing the layout.
- */
-function openstation_recycle_bin_render_template() {
-	ob_start();
-	?>
-	<div class="desktop-mode-recycle-bin" data-os-recycle-bin-root>
-		<header class="os-recycle-bin__toolbar" data-os-recycle-bin-toolbar>
-			<div class="os-recycle-bin__toolbar-left">
-				<os-segmented data-os-recycle-bin-filter>
-					<os-segment value="" selected><?php esc_html_e( 'All', 'desktop-mode' ); ?></os-segment>
-					<os-segment value="post"><?php esc_html_e( 'Posts', 'desktop-mode' ); ?></os-segment>
-					<os-segment value="page"><?php esc_html_e( 'Pages', 'desktop-mode' ); ?></os-segment>
-					<?php
-					// The Media segment is only useful when WP itself routes
-					// attachment deletions through Trash. That gate is the
-					// `MEDIA_TRASH` constant — defaults to false in core, can
-					// be flipped to true from `wp-config.php`. Without it,
-					// attachments permanent-delete on first click and the
-					// Trash bin will never have anything in this bucket, so
-					// the tab would always read "0" and confuse users.
-					if ( defined( 'MEDIA_TRASH' ) && MEDIA_TRASH ) :
-						?>
-						<os-segment value="attachment"><?php esc_html_e( 'Media', 'desktop-mode' ); ?></os-segment>
-						<?php
-					endif;
-					?>
-					<os-segment value="comment"><?php esc_html_e( 'Comments', 'desktop-mode' ); ?></os-segment>
-					<os-segment value="desktop"><?php esc_html_e( 'Desktop', 'desktop-mode' ); ?></os-segment>
-				</os-segmented>
-				<os-text-field
-					data-os-recycle-bin-search
-					placeholder="<?php esc_attr_e( 'Search trash…', 'desktop-mode' ); ?>"
-				></os-text-field>
-			</div>
-			<div class="os-recycle-bin__toolbar-right" data-os-recycle-bin-bulk hidden>
-				<span class="os-recycle-bin__count" data-os-recycle-bin-count></span>
-				<os-button variant="secondary" data-os-recycle-bin-restore-selected>
-					<span class="dashicons dashicons-image-rotate" aria-hidden="true"></span>
-					<?php esc_html_e( 'Restore', 'desktop-mode' ); ?>
-				</os-button>
-				<os-button variant="secondary" data-os-recycle-bin-pin-to-desktop>
-					<span class="dashicons dashicons-desktop" aria-hidden="true"></span>
-					<?php esc_html_e( 'Pin to desktop', 'desktop-mode' ); ?>
-				</os-button>
-				<os-button variant="danger" data-os-recycle-bin-purge-selected>
-					<span class="dashicons dashicons-trash" aria-hidden="true"></span>
-					<?php esc_html_e( 'Delete forever', 'desktop-mode' ); ?>
-				</os-button>
-			</div>
-			<div class="os-recycle-bin__toolbar-trailing">
-				<os-button variant="ghost" data-os-recycle-bin-refresh title="<?php esc_attr_e( 'Refresh', 'desktop-mode' ); ?>">
-					<span class="dashicons dashicons-update" aria-hidden="true"></span>
-				</os-button>
-				<os-button variant="danger" data-os-recycle-bin-empty>
-					<span class="dashicons dashicons-trash" aria-hidden="true"></span>
-					<?php esc_html_e( 'Empty Trash', 'desktop-mode' ); ?>
-				</os-button>
-			</div>
-		</header>
-		<div class="os-recycle-bin__body" data-os-recycle-bin-body>
-			<os-empty-state
-				data-os-recycle-bin-empty-state
-				role="status"
-				icon="trash"
-				heading="<?php esc_attr_e( 'The Trash is empty.', 'desktop-mode' ); ?>"
-				description="<?php esc_attr_e( 'Deleted posts, pages, and media show up here. Restoring puts them back where they were.', 'desktop-mode' ); ?>"
-				hidden
-			></os-empty-state>
-			<os-table
-				data-os-recycle-bin-table
-				selectable="multi"
-				sticky-header
-				hover
-				striped
-				loading
-				empty="<?php esc_attr_e( 'No items match the current filter or search.', 'desktop-mode' ); ?>"
-			></os-table>
-		</div>
-	</div>
-	<?php
-	$html = (string) ob_get_clean();
-
-	/**
-	 * Filter the recycle bin window's template HTML.
-	 *
-	 * Keep the `data-os-recycle-bin-*` hooks intact so the JS render
-	 * callback can find its mount points, or rename them and update the
-	 * matching constants in `src/recycle-bin/index.ts`.
-	 *
-	 * @param string $html Default template HTML.
-	 */
-	$filtered = (string) apply_filters( 'openstation_recycle_bin_template_html', $html );
-	echo wp_kses( $filtered, openstation_native_window_allowed_html() );
-}
-
-/**
  * Whether the current user should see the recycle bin at all.
  *
  * Filterable so plugins can hide it from authors/contributors who
  * don't manage trash, or invert the gate to expose it to a custom
- * role.
+ * role. The app's gate (`apps/trash/trash.os.php`) delegates here.
  *
  * @return bool
  */
@@ -251,91 +143,20 @@ function openstation_recycle_bin_user_can_use() {
 }
 
 /**
- * Register the recycle bin window on `init`.
+ * Enqueue the bin stylesheet with the shell.
  *
- * Hooked at priority 20, after `components.php` has bootstrapped the
- * native-window registry — same timing as the code editor.
+ * Not a window-open companion: the sheet also paints the
+ * drag-to-trash drop-target highlight on the CLOSED bin's dock tile
+ * (`data-os-trash-drop-active`), which must work before the window
+ * has ever opened.
  */
-function openstation_recycle_bin_register_window() {
+function openstation_recycle_bin_enqueue_style() {
 	if ( ! openstation_recycle_bin_user_can_use() ) {
 		return;
 	}
-
-	$icon_uris = openstation_recycle_bin_icon_uris();
-	$icon_uri  = $icon_uris['empty'];
-
-	$window_args = array(
-		'title'      => __( 'Trash', 'desktop-mode' ),
-		'icon'       => $icon_uri,
-		'template'   => 'openstation_recycle_bin_render_template',
-		'script'     => 'desktop-mode-recycle-bin',
-		'width'      => 880,
-		'height'     => 560,
-		'min_width'  => 520,
-		'min_height' => 360,
-		'placement'  => 'dock',
-		'nav_kind'   => 'control',
-		// Last on the rail, after the shell's own cluster (Mio 10,
-		// Overview 20, System 30). Trash is where things END UP, and a
-		// dock reads left to right: putting it anywhere but the end
-		// makes it one more app rather than the bottom of the pile.
-		'dock_order' => 40,
-		// The bin is the one dock tile a user can reasonably not want,
-		// so it gets a row in Apps & Plugins: dock (the default),
-		// desktop, both, or hidden. It registers no desktop icon, so
-		// that row is its only control.
-		'placeable'  => true,
-	);
-
-	/**
-	 * Filter the args used to register the recycle bin native window.
-	 *
-	 * @param array $window_args Args passed to `openstation_register_window()`.
-	 */
-	$window_args = (array) apply_filters( 'openstation_recycle_bin_window_args', $window_args );
-
-	$registered = openstation_register_window( 'desktop-mode-recycle-bin', $window_args );
-	if ( is_wp_error( $registered ) ) {
-		// phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
-		error_log( '[openstation] Recycle bin window registration failed: ' . $registered->get_error_message() );
-	}
-}
-add_action( 'init', 'openstation_recycle_bin_register_window', 20 );
-
-/**
- * Localize REST endpoints for the JS bundle.
- *
- * Same pattern as the code editor: the bundle reads its config off
- * `window.openStationRecycleBinConfig` and never hardcodes URLs.
- */
-function openstation_recycle_bin_localize_config() {
-	if ( ! openstation_recycle_bin_user_can_use() ) {
-		return;
-	}
-
-	wp_localize_script(
-		'desktop-mode-recycle-bin',
-		'openStationRecycleBinConfig',
-		array(
-			'restNonce'  => wp_create_nonce( 'wp_rest' ),
-			'listUrl'    => esc_url_raw( rest_url( 'desktop-mode/v1/recycle-bin' ) ),
-			'restoreUrl' => esc_url_raw( rest_url( 'desktop-mode/v1/recycle-bin/restore' ) ),
-			'purgeUrl'   => esc_url_raw( rest_url( 'desktop-mode/v1/recycle-bin/purge' ) ),
-			'emptyUrl'   => esc_url_raw( rest_url( 'desktop-mode/v1/recycle-bin/empty' ) ),
-			'countUrl'   => esc_url_raw( rest_url( 'desktop-mode/v1/recycle-bin/count' ) ),
-			'postTypes'  => openstation_recycle_bin_capture_post_types(),
-		)
-	);
-
 	wp_enqueue_style( 'desktop-mode-recycle-bin' );
 }
-// Priority 5, not an afterthought: `openstation_enqueue_assets()` (default 10)
-// harvests every lazy window's `wp_localize_script` data into the shell
-// payload, so config attached after 10 ships the bundle with no config — the
-// exact "openStationRecycleBinConfig is missing" failure the bundle warns
-// about. This ran at 30 and got away with it only while the bundle was
-// enqueued eagerly and WordPress printed the data itself at print time.
-add_action( 'admin_enqueue_scripts', 'openstation_recycle_bin_localize_config', 5 );
+add_action( 'admin_enqueue_scripts', 'openstation_recycle_bin_enqueue_style', 5 );
 
 /**
  * Inject the initial trash count and both bin drawings into the
